Switch to BigInteger for BspTree node indices Allows depth to exceed 31

+++ source/UnaryHeap/UnaryHeap.Algorithms/Spatial/Spatial2D.cs
@@ -47,7 +47,7 @@ namespace UnaryHeap.Algorithms
             /// <param name="result">The current list of interior leaves.</param>
             /// <param name="leafCount">The number of leaves in the tree.</param>
             public void InsideFilled(Point2D interiorPoint,
-                HashSet<int> result, int leafCount)
+                HashSet<BigInteger> result, int leafCount)
             {
 
             }
